Trekker ut VedtakTilsynBarn til egen fil
